Update README.md;Fixed Some Compilation Warnings

+++ HPL2/core/sources/gui/GuiPopUpFilePicker.cpp
@@ -229,7 +229,8 @@ namespace hpl {
 		
 		////////////////////////////////////////
         // If the selection is a system item (e.g. begins with "<") handle it differently
-        if(lSelection < (lNumItems - mvCurrentDirFullPath.size())) {
+        if(lSelection < ( lNumItems - static_cast<int>(mvCurrentDirFullPath.size()) )) 
+		{
             mvCurrentDirFullPath.clear();
             NavigateTo(pBox->GetItemText(lSelection));
         }
